Se modulariza parte del Home
