+++ src/components/MainLayout/MainLayout.jsx
@@ -1,0 +1,67 @@
+import React from 'react';
+
+import Button from '@material-ui/core/Button';
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import Grid from '@material-ui/core/Grid';
+import  { useStyles }  from './StyleMainLayout';
+import TopBar from '../TopBar';
+import Footer from '../Footer';
+import Container from '@material-ui/core/Container';
+import Typography from '@material-ui/core/Typography';
+
+const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9,10, 11, 12];
+
+export default function Album() {
+  const classes = useStyles();
+
+  return ( 
+    <>
+      <TopBar />
+      <main>
+        <div className={classes.heroContent}>
+          <Container maxWidth="sm">
+            <Typography component="h1" variant="h2" align="center" color="textPrimary" gutterBottom>
+                Aca va el banner
+            </Typography>
+          </Container>
+        </div>
+        <Container className={classes.cardGrid} maxWidth="lg">
+          <Grid container spacing={6}>
+            {cards.map((card) => (
+              <Grid item key={card} xs={12} sm={6} md={3}>
+                <Card className={classes.card}>
+                  <CardMedia
+                    className={classes.cardMedia}
+                    image="https://source.unsplash.com/random"
+                    title="Image title"
+                  />
+                  <CardContent className={classes.cardContent}>
+                    <Typography gutterBottom variant="h5" component="h2">
+                      Heading {card} 
+                    </Typography>
+                    <Typography>
+                      Esta es la descripcioin del item {card}
+                    </Typography>
+                  </CardContent>
+                  <CardActions>
+                    <Button size="small" color="primary">
+                      View
+                    </Button>
+                    <Button size="small" color="primary">
+                      Edit
+                    </Button>
+                  </CardActions>
+                </Card>
+              </Grid>
+            ))}
+          </Grid>
+        </Container>
+      </main>
+
+      <Footer/>
+    </>            
+  );
+}
