+++ src/components/TopBar/TopBar.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+
+const TopBar = () =>{
+    return (
+        <>
+            <CssBaseline />
+            <AppBar position="relative">
+                <Toolbar>
+                <Typography variant="h6" color="inherit" noWrap>
+                    Ventas - App
+                </Typography>
+                </Toolbar>
+            </AppBar>
+        </>
+    )    
+
+}
+export default TopBar;
