+++ src/components/TopBar/index.js
@@ -1,1 +1,1 @@
-export {default} from "./TopBar";
+export {default} from "./TopBar.jsx";
