--- src/layouts/MainLayout/MainLayout.scss
@@ -1,30 +1,0 @@
-@import "../../scss/index.scss";
-
-.main-layout {
-    background-color: $background-primary;
-    min-height: 100vh;
-    margin: 0;
-    
-
-    .top-bar {
-        height: 100px;
-    }
-
-    .content {
-        display: flex;
-        justify-content: center;
-
-        .menu-left {
-            width: 200px;
-            height: 100%;
-            margin-right: 10px;
-            margin-top: 10px;
-        }
-
-        .contentCenter {
-            width: 600px;
-            height: 100%;
-            margin-top: 10px;
-        }
-    }
-}
