--- src/layouts/MainLayout/index.js
@@ -1,1 +1,0 @@
-export {default} from "./MainLayout";
